-Items are now valued between 0 and 100 rather than 0 and 1.

+++ trunk/TableDraftingCode/Item.java
@@ -14,13 +14,13 @@ public class Item
     /**
      * How much this item is worth to each player
      */
-    public double[] m_values;
+    public int[] m_values;
 
     /**
      * Constructor
      * @param values The value of this item to each player
      */
-    public Item(double[] values)
+    public Item(int[] values)
     {
         m_values = values;
     }
